+++ mkdocs/changelog.md
@@ -1,5 +1,13 @@
+### Version 2.3 (11/23/2022)
 
-Version 2.2 (04/20/2022)
---------------------------------------------------------------------------------
+  - If the version of Pandoc is not supported, dont throw an error but warn,
+    and proceed as if it was the latest supported version (with undefined 
+    results if the two document models differ). 
+
+  - Add support for Pandoc 2.19, 2.19.1 and 2.19.2.
+
+### Version 2.2 (04/20/2022)
+
 
   - Add support for Pandoc 2.18.
+